Moved ParserConfig to config.py
